Moved createAccountsForProducer to managementsystem
Mkaes more sense to put it there
