+++ src/main/java/sdu/se9/tv2/management/system/persistence/IPersistenceAccount.java
@@ -50,14 +50,6 @@ public interface IPersistenceAccount {
     ProducerAccount createProducerAccount (String username, String password, int producerId) throws UsernameAlreadyExistsException;
 
     /**
-     * Adding accounts for a producer
-     * @param producerId
-     * @param amount
-     * @return
-     */
-    ArrayList<ProducerAccount> createAccountsForProducer (int producerId, int amount);
-
-    /**
      * Creates a new system admin account
      * @param username The username of the account
      * @param password The password of the account
